fix(219): surface activity as a primary section on the Dashboard
After the operator-dashboard reshuffle (PR #192), the activity surface
lived only at /overview/activity, reachable solely via the LiveNowBand
"View activity" link — which also renders on /operator (Settings). An
operator who runs `jinn run` and lands on /overview saw no activity feed.

The Dashboard answers "is anything happening right now?" and the activity
feed is exactly that answer. This makes it a primary section on /overview.

- Extract the In-flight + Recent activity cards from OverviewActivityPage
  into a reusable `ActivitySections` component (placement change, not a
  rebuild — same markup, test ids, and `['status']` query key).
- Render `<ActivitySections />` inline as a primary section on /overview.
- OverviewActivityPage keeps the dedicated /overview/activity drilldown,
  now composed from the shared component.
- Reword the recent empty state to "No recent activity yet." per the
  issue's "no recent activity" card-state requirement.
- SPA tests cover the new placement on /overview (loading, empty, error,
  in-flight, recent rows), the unit directly (ActivitySections.test.tsx),
  and the absence from /operator at the route level.

Closes #219
